requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
Django==4.2
django-anymail==9.1
django-ckeditor==6.5.1
django-cloudinary-storage==0.3.0
django-js-asset==2.0.0
django-ninja==0.21.0
idna==3.4
Pillow==9.5.0
pydantic==1.10.7
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
